% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze_thaw_count.R
\name{freeze.thaw.count}
\alias{freeze.thaw.count}
\title{Freeze/Thaw cycle counting function}
\usage{
freeze.thaw.count(temp.vec, time.vec, temp.threshold = 0, duration = 1, magnitudemin = 0, magnitudemax = 0)
}
\arguments{
\item{temp.vec}{A numeric vector of consecutive temperature readings in which to search for freeze/thaw cycles.}

\item{time.vec}{A vector of consecutive date/time data corresponding with the temperature data provided to temp.vec. Importantly, the function assumes this data has no missing data internally and that the time steps remain uniform throughout (e.g., the data were recorded every 30 minutes). If either of these assumptions are not satisfied, this function will produce erroneous results.}

\item{temp.threshold}{A numeric value of length 1 corresponding to the maximum temperature at which the medium in question is considered "frozen." Freeze/thaw cycles will be assumed to be fluctuations about this temperature. Assumes the temperature data are in Celcius and thus defaults to 0 degrees.}

\item{duration}{A numeric value of length 1 corresponding to the minimum number of time steps the medium in question must be both "frozen" and "thawed" for in order for a freeze/thaw cycle to have occurred. Defaults to 1 time step.}

\item{magnitudemin, magnitudemax}{Numeric values each of length 1 corresponding to the extent to which temperature must fall below (magnitudemin) and/or rise above (magnitude max) the temperature threshold set for the medium in question to be considered "thawed" or "frozen" enough for one's purposes. In other words, the function will not begin checking against the duration length until temperatures reach values beyond those specified by the magnitudemin and magnitudemax arguments. Both default to 0, indicating that any departure above or below the threshold value is sufficient. Note: Both arguments are to be given as absolute values and should be positive.}
}
\description{
This function is designed to take two vectors--a vector of continuous temperature readings and a vector of continuous date/time information--and count the number of freeze/thaw cycles that occurred over the time period in question. Purposefully, the function is designed to allow for adjustment of the parameters that may define a freeze/thaw cycle. temp.vec and time.vec should usually be from the same data source, but the function will run properly so long as they are equal in length.
}
\examples{
freeze.thaw.count()
}
\keyword{cycles}
\keyword{freeze}
\keyword{temperature}
\keyword{thaw}
\keyword{time}
