% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze_thaw_sweep.R
\name{freeze.thaw.sweep}
\alias{freeze.thaw.sweep}
\title{Freeze/Thaw cycle parameter sweeping function}
\usage{
freeze.thaw.sweep(durations = 1, magnitudemins = 0, thresholds = 0, temp.vec, time.vec)
}
\arguments{
\item{temp.vec}{A numeric vector of consecutive temperature readings in which to search for freeze/thaw cycles.}

\item{time.vec}{A vector of consecutive date/time data corresponding with the temperature data provided to temp.vec. Importantly, the function assumes this data has no missing data internally and that the time steps remain uniform throughout (e.g., the data were recorded every 30 minutes). If either of these assumptions are not satisfied, this function will produce erroneous results.}

\item{durations}{A numeric vector length 1 or greater corresponding to the minimum numbers of time steps the medium in question must be both "frozen" and "thawed" for in order for a freeze/thaw cycle to have occurred. Defaults to 1 time step.}

\item{magnitudemins}{A numeric vector of length 1 corresponding to the extents to which temperature must fall below (magnitudemin) the temperature threshold set for the medium in question to be considered "frozen" enough for one's purposes. In other words, the function will not begin checking against the duration length until temperatures reach values fall below those specified by this argument, relative to the threshold. Defaults to 0, indicating that any detectable departure below the threshold value is sufficient. Note: Argument values are to be given as absolute values and should thus be positive.}

\item{thresholds}{A numeric vector of length 1 or greater corresponding to the maximum temperatures at which the medium in question is to be considered "frozen." Freeze/thaw cycles will be assumed to be fluctuations about these temperatures. Assumes the temperature data are in Celcius and defaults to 0 degrees.}
}
\description{
This function is designed to take two vectors--a vector of continuous temperature readings and a vector of continuous date/time information--and count the number of freeze/thaw cycles that occurred over the time period in question under a range of operational definitions of a freeze/thaw cycle. Specifically, the function allows manipulation of the duration, temp.threshold, and magnitudemin arguments of the freeze.thaw.count function. See the documentation for that function for details.
}
\examples{
freeze.thaw.sweep()
}
\keyword{cycles}
\keyword{freeze}
\keyword{parameter}
\keyword{sweep}
\keyword{temperature}
\keyword{thaw}
\keyword{time}
