% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze_thaw_count.R
\name{freeze.thaw.count}
\alias{freeze.thaw.count}
\title{Basic freeze-thaw cycle (FTC) counting function}
\usage{
freeze.thaw.count(temp.vec, time.vec, temp.threshold = 0, duration = 1, magnitudemin = 0, magnitudemax = 0)
}
\arguments{
\item{temp.vec}{A numeric vector of consecutive temperature readings in which to search for freeze-thaw cycles. It is assumed that the data are continuous and missing no data. If either assumption is violated, the output data will be inaccurate!}

\item{time.vec}{A vector of consecutive data-time data corresponding with the temperature data provided to temp.vec. Importantly, the function assumes this data has no missing data internally and that the time steps remain uniform throughout (e.g., the data were recorded every 30 minutes without fail). If either of these assumptions are not satisfied, this function will produce erroneous results, as noted above.}

\item{temp.threshold}{A numeric value of length 1 corresponding to the maximum temperature at which the medium in question is considered "frozen." Freeze-thaw cycles will be assumed to be fluctuations about this temperature at a minimum. Defaults to 0  under the assumption the data given to temp.vec are in degrees Celcius but can be switched to any value to accomodate other temperature scales.}

\item{duration}{A numeric value of length 1 corresponding to the minimum number of time steps the medium in question must be both "frozen" and "thawed" for a complete freeze-thaw cycle to have occurred. Defaults to 1, as in one time step.}

\item{magnitudemin, magnitudemax}{Numeric values each of length 1 corresponding to the extent to which the temperature must fall below (magnitudemin) and/or rise above (magnitudemax) the temperature threshold set for the medium in question to be considered "effectively thawed" or "effective frozen." In other words, the function will not begin checking against the duration length until temperatures reach values beyond those specified by the magnitudemin and magnitudemax arguments (in combination with the threshold argument). Both default to 0, indicating that any departure above or below the threshold value is sufficient. Note: Both arguments are to be given as absolute values (deviations from the temperature threshold) and thus should be specificed as positive values.}
}
\description{
This function is designed to take two vectors--a vector of continuous temperature readings and a vector of continuous date-time data--and count the number of freeze-thaw cycles (FTCs) that occurred over the given time period. Purposefully, the function is designed to allow for adjustment of the criteria that define a freeze-thaw cycle. temp.vec and time.vec should usually be from the same data source, but the function will run so long as they are equal in length.
}
\examples{
freeze.thaw.count()
}
\keyword{cycles}
\keyword{freeze}
\keyword{temperature}
\keyword{thaw}
\keyword{time}
